requirements.txt
openai>=1.30.0
streamlit>=1.35.0
numpy>=1.24.0
pandas>=2.0.0
tiktoken>=0.7.0
python-dotenv>=1.0.0
